# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Nexus Supply Chain API
  version: 1.0.0
  description: Inventory, purchase orders, user administration, notifications, and audit records. Bearer
    authentication is required unless an operation declares otherwise. Endpoint roles are documented per
    operation; there is no per-warehouse tenancy model.
servers:
- url: http://localhost:8080/api/v1
  description: Local development
security:
- BearerAuth: []
paths:
...
  /orders:
    get:
      summary: List orders
      description: Staff and administrators can view orders across the shared workspace. Newest first;
        at most 50 per page.
      operationId: getOrders
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
    post:
      summary: Create a purchase order
      description: Staff and administrators. Creates a DRAFT using product prices from the database. Every
        product must be active and assigned to the destination warehouse.
      operationId: postOrders
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
  /orders/{id}:
    get:
      summary: Get an order
      description: Staff and administrators.
      operationId: getOrdersId
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      parameters:
      - $ref: '#/components/parameters/Id'
  /orders/{id}/status:
    put:
      summary: Update an order status
      description: Staff and administrators. Forward transitions are DRAFT → PENDING_APPROVAL → APPROVED
        → SHIPPED → DELIVERED; any nonterminal state can transition to CANCELLED. Only administrators
        can move to APPROVED, SHIPPED, or DELIVERED. Staff can submit drafts or cancel orders only while
        DRAFT or PENDING_APPROVAL. Repeating the current status changes nothing. Delivery adds stock exactly
        once.
      operationId: putOrdersIdStatus
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdateRequest'
      parameters:
      - $ref: '#/components/parameters/Id'
  /categories:
    get:
      summary: List categories
      description: Staff and administrators.
      operationId: getCategories
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    post:
      summary: Create a category
      description: Administrators only.
      operationId: postCategories
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
  /categories/{id}:
    get:
      summary: Get a category
      description: Staff and administrators.
      operationId: getCategoriesId
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      parameters:
      - $ref: '#/components/parameters/Id'
    put:
      summary: Update a category
      description: Administrators only. Replaces the editable fields; omitted optional fields become null.
      operationId: putCategoriesId
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      summary: Delete a category
      description: Administrators only. Database references may prevent deletion.
      operationId: deleteCategoriesId
      responses:
        '204':
          description: Successful response
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
      parameters:
      - $ref: '#/components/parameters/Id'
...
  /users:
    get:
      summary: List users
      description: Administrators only. Password hashes are excluded.
      operationId: getUsers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    post:
      summary: Create a user
      description: Administrators only. Creates an ACTIVE account with the selected role.
      operationId: postUsers
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
      description: Zero-based page number.
    Size:
      name: size
      in: query
      schema:
        type: integer
        format: int32
        default: 50
        minimum: 1
      description: Requested page size. Values above 50 are capped at 50.
  schemas:
...
    OrderCreateRequest:
      type: object
      required:
      - supplierId
      - warehouseId
      - items
      properties:
        supplierId:
          type: string
          format: uuid
          description: Must identify an active supplier.
        warehouseId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          minItems: 1
          maxItems: 500
          description: Each product may appear only once, must be active, and must be assigned to the
            order warehouse. Prices come from the server. The order total cannot exceed 9999999999.99.
    OrderItemRequest:
      type: object
      required:
      - productId
      - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
    OrderStatus:
      type: string
      enum:
      - DRAFT
      - PENDING_APPROVAL
      - APPROVED
      - SHIPPED
      - DELIVERED
      - CANCELLED
    OrderStatusUpdateRequest:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
      description: Status names are case-insensitive. Repeating the current status is an idempotent no-op;
        other updates must follow the documented transition rules.
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        subtotal:
          type: number
        productName:
          type: string
        productSku:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        supplierId:
          type: string
          format: uuid
        supplierName:
          type: string
        warehouseId:
          type: string
          format: uuid
        warehouseName:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
        createdBy:
          type: string
          format: uuid
          nullable: true
          description: Null for automatically created orders.
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
...
    OrderPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        totalElements:
          type: integer
          format: int64
          description: 'Always -1: the service uses slice pagination without a total count.'
          enum:
          - -1
        totalPages:
          type: integer
          format: int32
          description: 'Always -1: use hasNext to navigate.'
          enum:
          - -1
        pageNumber:
          type: integer
          format: int32
          minimum: 0
        pageSize:
          type: integer
          format: int32
          minimum: 1
          maximum: 50
        hasNext:
          type: boolean
    CategoryRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 255
          nullable: true
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
...
    UserCreateRequest:
      type: object
      required:
      - fullName
      - email
      - password
      - roleName
      properties:
        fullName:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          writeOnly: true
          description: At most 72 UTF-8 bytes.
        roleName:
          type: string
          enum:
          - ROLE_ADMIN
          - ROLE_STAFF
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
          - ROLE_ADMIN
          - ROLE_STAFF
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
  responses:
    '400':
      description: Invalid request or disallowed state transition. Malformed transport input may use Spring’s
        standard error body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    '401':
      description: Missing, expired, invalid, or revoked credentials. Login returns ApiError; authentication
        filters may return an empty or framework error body.
    '403':
      description: Authenticated user lacks permission. Service errors use ApiError; security-filter rejections
        may use a framework error body.
    '404':
      description: Referenced resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'